# internal/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Docveta API
  version: '1'
  description: 'HTTP API of Docveta, a self-hosted document manager.


    Errors are RFC 9457 problem details (`application/problem+json`). Cookie-authenticated unsafe requests must be same-origin
    (CSRF protection); bearer tokens are exempt.


    This file is the contract: `go test ./internal/api` fails when a route is added or removed without updating it, and the
    integration test validates responses against it.'
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
- url: /
security:
- sessionCookie: []
- bearerToken: []
...
paths:
...
  /mcp:
    post:
      summary: Model Context Protocol endpoint (JSON-RPC 2.0); authenticate with a personal access token
      tags:
      - ai
      description: 'Tools: search_documents, get_document, get_document_text, find_similar_documents, list_spaces, list_tags,
        list_correspondents, list_document_types.'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        default:
          $ref: '#/components/responses/Problem'
        '202':
          description: A notification was accepted (no answer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RPCRequest'
    get:
      summary: 'Not supported (405): there is no server-initiated stream'
      tags:
      - ai
      responses:
        '405':
          description: Method not allowed
          content:
            text/plain:
              schema:
                type: string
        default:
          $ref: '#/components/responses/Problem'
...
  /api/v1/uploads:
    options:
      summary: 'tus: what the server supports'
      tags:
      - documents
      description: Resumable uploads follow the tus 1.0.0 protocol (creation, termination, expiration).
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        default:
          $ref: '#/components/responses/Problem'
      security: []
    post:
      summary: 'tus: start a resumable upload'
      tags:
      - documents
      description: 'Metadata keys (base64 values): space_id, filename, title, tag_ids, correspondent_id, document_type_id,
        document_date, language, allow_duplicate, source.'
      responses:
        '201':
          description: Created; the Location header is the upload URL
        default:
          $ref: '#/components/responses/Problem'
      parameters:
      - name: Tus-Resumable
        in: header
        schema:
          type: string
        required: true
      - name: Upload-Length
        in: header
        schema:
          type: integer
        required: true
      - name: Upload-Metadata
        in: header
        schema:
          type: string
  /api/v1/uploads/{id}:
    head:
      summary: 'tus: how much of the upload has arrived (Upload-Offset)'
      tags:
      - documents
      responses:
        '200':
          description: OK
        default:
          $ref: '#/components/responses/Problem'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: Tus-Resumable
        in: header
        schema:
          type: string
        required: true
    patch:
      summary: 'tus: send the next chunk'
      tags:
      - documents
      description: When the last chunk arrives the file becomes a document; its id is in the Docveta-Document-Id response
        header.
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        default:
          $ref: '#/components/responses/Problem'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: Tus-Resumable
        in: header
        schema:
          type: string
        required: true
      - name: Upload-Offset
        in: header
        schema:
          type: integer
        required: true
      requestBody:
        required: true
        content:
          application/offset+octet-stream:
            schema:
              type: string
              format: binary
    delete:
      summary: 'tus: cancel an upload'
      tags:
      - documents
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        default:
          $ref: '#/components/responses/Problem'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: Tus-Resumable
        in: header
        schema:
          type: string
        required: true
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-docveta_session
      description: Set by login/setup. Named `docveta_session` when the base URL is plain http.
    bearerToken:
      type: http
      scheme: bearer
      description: Personal access token (`dvt_pat_…`) created in Settings → API tokens.
    workerToken:
      type: http
      scheme: bearer
      description: Worker token (`dvt_wrk_…`) from Administration → Processing.
  responses:
    NoContent:
      description: Done
    Problem:
      description: Error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Problem:
      properties:
        code:
          type: string
        detail:
          type: string
        errors:
          items:
            properties:
              field:
                type: string
              message:
                type: string
            required:
            - field
            - message
            type: object
          type: array
        extra:
          additionalProperties:
            nullable: true
          type: object
        instance:
          type: string
        status:
          type: integer
        title:
          type: string
        type:
          type: string
      required:
      - type
      - title
      - status
      - code
      type: object
...
    RPCRequest:
      type: object
      properties:
        jsonrpc:
          type: string
        id:
          nullable: true
        method:
          type: string
        params:
          nullable: true
...